requirements.txt
# Smart-Gemini PDC - Dependências Python

# Framework Web
fastapi==0.109.0
uvicorn[standard]==0.27.0
jinja2==3.1.3
python-multipart==0.0.6

# Banco de Dados
sqlalchemy==2.0.25
#psycopg2-binary==2.9.11  # Para PostgreSQL/TimescaleDB

# Validação e Configuração
#pydantic
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Redis
redis==5.0.1

# Processamento de Dados
numpy==1.26.3

# Gerenciamento de Processos
psutil==5.9.8

# Protocolo IEEE C37.118
crcmod==1.7

# Logging (já incluído no Python, mas pode especificar versões específicas se necessário)

# Desenvolvimento (opcional)
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
flake8==7.0.0